Adicionando

+++ S13-Projeto-3-CRUD-BD/src/controller/coachController.js
@@ -1,5 +1,4 @@
-
-// post -> criar um treinador
+//post -> criar um treinador
 // get -> visualizar todos os treinadores
 // get -> visualizar um treinador por id
 // patch -> 
@@ -13,8 +12,25 @@
  * const team = req.body.team
  */
 const CoachModel = require('../models/coachModel')
+const jwt = require('jsonwebtoken')
+const SECRET = process.env.SECRET
+
 const createCoach = async (req, res) => {
    try {
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Cadê o authorization do role?')
+    }
+
+    const token = authHeader.split(' ') [1]
+
+    await jwt.verify(token, SECRET, async function (erro) {
+
+      if (erro) {
+        return res.status(403).send('Não vai rolar')
+    }
       const { name, team, region, age, gender } = req.body
 
       const newCoach = new CoachModel({
@@ -24,6 +40,7 @@ const createCoach = async (req, res) => {
       const savedCoach = await newCoach.save()
 
       res.status(201).json(savedCoach)
+    })
    } catch (error) {
      console.error(error)
      res.status(500).json({ message: error.message })
@@ -32,8 +49,25 @@ const createCoach = async (req, res) => {
 
 const findAllCoaches = async (req, res) => {
   try {
-    const allCoaches = await CoachModel.find()
-    res.status(200).json(allCoaches)
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Cadê o authorization do role?')
+    }
+
+    const token = authHeader.split(' ') [1]
+
+    await jwt.verify(token, SECRET, async function (erro) {
+
+      if (erro) {
+        return res.status(403).send('Não vai rolar')
+      }
+
+      const allCoaches = await CoachModel.find()
+      res.status(200).json(allCoaches)
+    })
+
   } catch(error) {
     console.error(error)
     res.status(500).json({ message: error.message})
@@ -42,8 +76,24 @@ const findAllCoaches = async (req, res) => {
 
 const findCoachById = async (req, res) => {
    try {
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Cadê o authorization do role?')
+    }
+
+    const token = authHeader.split(' ') [1]
+
+    await jwt.verify(token, SECRET, async function (erro) {
+
+      if (erro) {
+        return res.status(403).send('Não vai rolar')
+    }
+
      const findCoach = await CoachModel.findById(req.params.id)
      res.status(200).json(findCoach)
+  })
    } catch (error) {
      console.error(error)
      res.status(500).json({ message: error.message })
@@ -52,12 +102,27 @@ const findCoachById = async (req, res) => {
 
 const updateCoach = async (req, res) => {
   try {
+    
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Cadê o authorization do role?')
+    }
+
+    const token = authHeader.split(' ') [1]
+
+    await jwt.verify(token, SECRET, async function (erro) {
+
+      if (erro) {
+        return res.status(403).send('Não vai rolar')
+    }
     const { name, age, region, team, gender } = req.body
     const updatedCoach = await CoachModel
     .findByIdAndUpdate(req.params.id, {
       name, age, region, team, gender
     })
     res.status(200).json(updatedCoach)
+    })
   } catch (error) {
     console.error(error)
     res.status(500).json({ message: error.message })
@@ -66,10 +131,25 @@ const updateCoach = async (req, res) => {
 
 const deleteCoach = async (req, res) => {
    try {
+
+    const authHeader = req.get('authorization')
+
+    if (!authHeader) {
+      return res.status(401).send('Cadê o authorization do role?')
+    }
+
+    const token = authHeader.split(' ') [1]
+
+    await jwt.verify(token, SECRET, async function (erro) {
+
+      if (erro) {
+        return res.status(403).send('Não vai rolar')
+    }
        const { id } = req.params
        await CoachModel.findByIdAndDelete(id)
        const message = `O treinador com o ${id} foi deletado com sucesso!`
       res.status(200).json({ message })
+  })
    } catch (error) {
      console.error(error)
      res.status(500).json({ message: error.message })
